feat: 선택 카드 내역 조회에 pageable 구현

+++ backend/src/main/java/CardRecommendService/cardHistory/FindAllResponse.java
@@ -3,6 +3,6 @@ package CardRecommendService.cardHistory;
 import java.util.List;
 
 public record FindAllResponse(List<CardHistoryResponse> cardHistoryList,
-                              int totalCount,
+                              Long totalCount,
                               Integer totalCost) {
 }
